pyproject.toml
[project]
name = "gong-to-github"
version = "0.1.0"
description = "Sync Gong call transcripts to GitHub as Markdown files"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "pygithub>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
gong-sync = "gong_to_github.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
